Completed fairness labelling functionality
Completed functionality to add a fairness label to each data instance. To achieve this, some decisions needed to be made on how to determine whether an instance passes or fails, and how to normalize group fairness values to be either positive or negative. These decisions should be reevaluated with the full team