Refactored mock data generation
Refactored the data generation code to minimize what each dataset specific script needs to include. Scripts for getting specific datasets (e.g. the COMPAS dataset) now only need to specify: Their expected input and output attributes, a list of sensitive attributes, the path to their dataset file, and a function for determining whether a given data instance from the dataset passes or fails (as though it were the input to a unit test).